Reposition and refactor common operations
