+++ react/src/utils/lectureUtils.js
@@ -154,178 +154,3 @@ export const getColorNumber = (lecture) => (
 export const getSyllabusUrl = (lecture) => (
   `https://cais.kaist.ac.kr/syllabusInfo?year=${lecture.year}&term=${lecture.semester}&subject_no=${lecture.code}&lecture_class=${lecture.class_no}&dept_id=${lecture.department}`
 );
-
-export const performAddToTable = (
-  caller,
-  lecture, selectedTimetable, user, fromString,
-  addLectureToTimetableDispatch
-) => {
-  if (
-    lecture.classtimes.some((ct1) => (
-      selectedTimetable.lectures.some((l) => (
-        l.classtimes.some((ct2) => (
-          (ct2.day === ct1.day)
-          && (ct2.begin < ct1.end)
-          && (ct2.end > ct1.begin)
-        ))
-      ))
-    ))
-  ) {
-    // eslint-disable-next-line no-alert
-    alert(i18n.t('ui.message.timetableOverlap'));
-    return;
-  }
-
-  if (!user) {
-    addLectureToTimetableDispatch(lecture);
-  }
-  else {
-    axios.post(
-      `/api/users/${user.id}/timetables/${selectedTimetable.id}/add-lecture`,
-      {
-        lecture: lecture.id,
-      },
-      {
-        metadata: {
-          gaCategory: 'Timetable',
-          gaVariable: 'POST Update / Instance',
-        },
-      },
-    )
-      .then((response) => {
-        const newProps = caller.props;
-        if (!newProps.selectedTimetable || newProps.selectedTimetable.id !== selectedTimetable.id) {
-          return;
-        }
-        // TODO: Fix timetable not updated when semester unchanged and timetable changed
-        addLectureToTimetableDispatch(lecture);
-      })
-      .catch((error) => {
-      });
-  }
-
-  ReactGA.event({
-    category: 'Timetable - Lecture',
-    action: 'Added Lecture to Timetable',
-    label: `Lecture : ${lecture.id} / From : ${fromString}`,
-  });
-};
-
-export const performDeleteFromTable = (
-  caller,
-  lecture, selectedTimetable, user, fromString,
-  removeLectureFromTimetableDispatch
-) => {
-  if (!user) {
-    removeLectureFromTimetableDispatch(lecture);
-  }
-  else {
-    axios.post(
-      `/api/users/${user.id}/timetables/${selectedTimetable.id}/remove-lecture`,
-      {
-        lecture: lecture.id,
-      },
-      {
-        metadata: {
-          gaCategory: 'Timetable',
-          gaVariable: 'POST Update / Instance',
-        },
-      },
-    )
-      .then((response) => {
-        const newProps = caller.props;
-        if (!newProps.selectedTimetable || newProps.selectedTimetable.id !== selectedTimetable.id) {
-          return;
-        }
-        // TODO: Fix timetable not updated when semester unchanged and timetable changed
-        removeLectureFromTimetableDispatch(lecture);
-      })
-      .catch((error) => {
-      });
-  }
-
-  ReactGA.event({
-    category: 'Timetable - Lecture',
-    action: 'Deleted Lecture from Timetable',
-    label: `Lecture : ${lecture.id} / From : ${fromString}`,
-  });
-};
-
-export const performAddToCart = (
-  caller,
-  lecture, year, semester, user, fromString,
-  addLectureToCartDispatch
-) => {
-  if (!user) {
-    addLectureToCartDispatch(lecture);
-  }
-  else {
-    axios.post(
-      `/api/users/${user.id}/wishlist/add-lecture`,
-      {
-        lecture: lecture.id,
-      },
-      {
-        metadata: {
-          gaCategory: 'Wishlist',
-          gaVariable: 'POST Update / Instance',
-        },
-      },
-    )
-      .then((response) => {
-        const newProps = caller.props;
-        if (newProps.year !== year || (newProps.semester !== semester)
-        ) {
-          return;
-        }
-        addLectureToCartDispatch(lecture);
-      })
-      .catch((error) => {
-      });
-  }
-
-  ReactGA.event({
-    category: 'Timetable - Lecture',
-    action: 'Added Lecture to Cart',
-    label: `Lecture : ${lecture.id} / From : ${fromString}`,
-  });
-};
-
-export const performDeleteFromCart = (
-  caller,
-  lecture, year, semester, user, fromString,
-  deleteLectureFromCartDispatch
-) => {
-  if (!user) {
-    deleteLectureFromCartDispatch(lecture);
-  }
-  else {
-    axios.post(
-      `/api/users/${user.id}/wishlist/remove-lecture`,
-      {
-        lecture: lecture.id,
-      },
-      {
-        metadata: {
-          gaCategory: 'Wishlist',
-          gaVariable: 'POST Update / Instance',
-        },
-      },
-    )
-      .then((response) => {
-        const newProps = caller.props;
-        if (newProps.year !== year || newProps.semester !== semester) {
-          return;
-        }
-        deleteLectureFromCartDispatch(lecture);
-      })
-      .catch((error) => {
-      });
-  }
-
-  ReactGA.event({
-    category: 'Timetable - Lecture',
-    action: 'Deleted Lecture from Cart',
-    label: `Lecture : ${lecture.id} / From : ${fromString}`,
-  });
-};
